Validation method is so tied to token that it would be good to keep it on te token model. Also fix two import bugs
